added a way to disable Santos widget

+++ .config/qtile/configs/widgets/santos.sh
@@ -1,8 +1,14 @@
 #!/bin/bash
 
 if [[ -f "/tmp/santosmatch" ]]; then
-  santos=$(</tmp/santosmatch)
-  output="  $santos"
-  printf '<span rise="4200" foreground="#EF5A6F">%s</span> <span size="x-large" foreground="#fff">|</span>' "$output"
+  if [[ ! -f "/tmp/stop_santos_widget" ]]; then
+    santos=$(</tmp/santosmatch)
+    output="  $santos"
+    printf '<span rise="4200" foreground="#EF5A6F">%s</span> <span size="x-large" foreground="#fff">|</span>' "$output"
+  fi
+else
+  if [[ -f "/tmp/stop_santos_widget" ]]; then
+    rm "/tmp/stop_santos_widget"
+  fi
 fi
